+++ src/modules/paperDiscovery/index.ts
@@ -2,3 +2,4 @@ export * from "./types";
 export * from "./providers";
 export * from "./search";
 export * from "./importer";
+export * from "./references";